+++ tests/Comparison/ArrayComparisonTest.php
@@ -4,6 +4,15 @@ namespace PhpAbac\Test\Comparison;
 
 use PhpAbac\Comparison\ArrayComparison;
 
+use PhpAbac\Manager\ComparisonManager;
+use PhpAbac\Manager\AttributeManager;
+
+use PhpAbac\Model\PolicyRuleAttribute;
+use PhpAbac\Model\Attribute;
+
+use PhpAbac\Example\User;
+use PhpAbac\Example\Visa;
+
 class ArrayComparisonTest extends \PHPUnit_Framework_TestCase
 {
     /** @var ArrayComparison **/
@@ -11,7 +20,7 @@ class ArrayComparisonTest extends \PHPUnit_Framework_TestCase
 
     public function setUp()
     {
-        $this->comparison = new ArrayComparison();
+        $this->comparison = new ArrayComparison(new ComparisonManager(new AttributeManager([])));
     }
 
     public function testIsIn()
@@ -76,5 +85,59 @@ class ArrayComparisonTest extends \PHPUnit_Framework_TestCase
             'ROLE_USER',
             'ROLE_POST_MANAGER',
         ]));
+    }
+    
+    public function testContains() {
+        $visas = include(__DIR__ . '/../fixtures/visas.php');
+        $policyRuleAttributes = [
+            (new PolicyRuleAttribute())
+            ->setAttribute(
+                (new Attribute())
+                ->setType('resource')
+                ->setProperty('country')
+                ->setName('Pays')
+                ->setSlug('pays')
+            )
+            ->setComparison('isEqual')
+            ->setComparisonType('string')
+            ->setValue('US'),
+            (new PolicyRuleAttribute())
+            ->setAttribute(
+                (new Attribute())
+                ->setType('resource')
+                ->setProperty('lastRenewal')
+                ->setName('Dernier renouvellement')
+                ->setSlug('dernier-renouvellement')
+            )
+            ->setComparison('isMoreRecentThan')
+            ->setComparisonType('datetime')
+            ->setValue('-1Y'),
+        ];
+        $extraData = [
+            'attribute' => 
+                (new Attribute())
+                ->setProperty('visas')
+                ->setName('Visas')
+                ->setSlug('visas')
+                ->setType('resource')
+                ->setValue([$visas[0], $visas[1]])
+            ,
+            'user' => 
+                (new User())
+                ->setId(1)
+                ->setName('John Doe')
+                ->setAge(36)
+                ->setParentNationality('FR')
+                ->addVisa($visas[0])
+                ->addVisa($visas[1])
+                ->setHasDoneJapd(true)
+                ->setHasDrivingLicense(false)
+            ,
+            'resource' => null
+        ];
+        $this->assertFalse($this->comparison->contains($policyRuleAttributes, [$visas[0], $visas[1]], $extraData));
+        $extraData['user']->addVisa($visas[2]);
+        $extraData['attribute']->setValue($visas);
+        $this->assertTrue($this->comparison->contains($policyRuleAttributes, [$visas[0], $visas[1], $visas[2]], $extraData));
     }
 }